Move tools key-generation-service imports
